crash_tracker: import limiter code
Shifting the implementation over to the crash tracker module. As part of
that the implementation is co-routinized, and the fields of the struct
crash_tracker_metadata are renamed to remove the `_` prefix.

Pure refactor, no behaviour change intended.
