+++ src/v/crash_tracker/BUILD
@@ -16,11 +16,20 @@ redpanda_cc_library(
         "types.h",
     ],
     implementation_deps = [
+        "//src/v/bytes",
+        "//src/v/bytes:iobuf",
+        "//src/v/config",
+        "//src/v/hashing:xx",
         "//src/v/model",
+        "//src/v/random:generators",
         "//src/v/serde",
         "//src/v/serde:bytes",
         "//src/v/serde:enum",
+        "//src/v/serde:iobuf",
         "//src/v/serde:sstring",
+        "//src/v/serde:vector",
+        "//src/v/utils:file_io",
+        "@fmt",
     ],
     include_prefix = "crash_tracker",
     visibility = ["//visibility:public"],
