+++ src/v/redpanda/application.cc
@@ -183,9 +183,6 @@
 #include <memory>
 #include <vector>
 
-// Crash tracking resets every 1h.
-static constexpr model::timestamp_clock::duration crash_reset_duration{1h};
-
 static void set_local_kafka_client_config(
   std::optional<kafka::client::configuration>& client_config,
   const config::node_config& config) {
@@ -502,7 +499,7 @@ int application::run(int ac, char** av) {
                 hydrate_config(cfg);
                 initialize();
                 check_environment();
-                check_for_crash_loop(app_signal.abort_source());
+                init_crashtracker(app_signal);
                 setup_metrics();
                 wire_up_and_start(app_signal);
                 post_start_tasks();
@@ -1024,110 +1021,9 @@ void application::check_environment() {
     }
 }
 
-/// Here we check for too many consecutive unclean shutdowns/crashes
-/// and abort the startup sequence if the limit exceeds
-/// crash_loop_limit until the operator intervenes. Crash tracking
-/// is reset if the node configuration changes or its been 1h since
-/// the broker last failed to start. This metadata is tracked in the
-/// tracker file. This is to prevent on disk state from piling up in
-/// each unclean run and creating more state to recover for the next run.
-void application::check_for_crash_loop(ss::abort_source& as) {
-    if (config::node().developer_mode()) {
-        // crash loop tracking has value only in long running clusters
-        // that can potentially accumulate state across restarts.
-        return;
-    }
-    auto file_path = config::node().crash_loop_tracker_path();
-    std::optional<crash_tracker_metadata> maybe_crash_md;
-    if (
-      // Tracking is reset every time the broker boots in recovery mode.
-      !config::node().recovery_mode_enabled()
-      && ss::file_exists(file_path.string()).get()) {
-        // Ok to read the entire file, it contains a serialized uint32_t.
-        auto buf = read_fully(file_path).get();
-        try {
-            maybe_crash_md = serde::from_iobuf<crash_tracker_metadata>(
-              std::move(buf));
-        } catch (const serde::serde_exception&) {
-            // A malformed log file, ignore and reset it later.
-            // We truncate it below.
-            vlog(_log.warn, "Ignorning malformed tracker file {}", file_path);
-        }
-    }
-
-    // Compute the checksum of the current node configuration.
-    auto current_config
-      = read_fully_to_string(config::node().get_cfg_file_path()).get();
-    auto checksum = xxhash_64(current_config.c_str(), current_config.length());
-
-    if (maybe_crash_md) {
-        auto& crash_md = maybe_crash_md.value();
-        auto& limit = config::node().crash_loop_limit.value();
-
-        // Check if it has been atleast 1h since last unsuccessful restart.
-        // Tracking resets every 1h.
-        auto time_since_last_start
-          = model::duration_since_epoch(model::timestamp::now())
-            - model::duration_since_epoch(crash_md._last_start_ts);
-
-        auto crash_limit_ok = !limit || crash_md._crash_count <= limit.value();
-        auto node_config_changed = crash_md._config_checksum != checksum;
-        auto tracking_reset = time_since_last_start > crash_reset_duration;
-
-        auto ok_to_proceed = crash_limit_ok || node_config_changed
-                             || tracking_reset;
-
-        if (!ok_to_proceed) {
-            vlog(
-              _log.error,
-              "Crash loop detected. Too many consecutive crashes {}, exceeded "
-              "{} configured value {}. To recover Redpanda from this state, "
-              "manually remove file at path {}. Crash loop automatically "
-              "resets 1h after last crash or with node configuration changes.",
-              crash_md._crash_count,
-              config::node().crash_loop_limit.name(),
-              limit.value(),
-              file_path);
-
-            const auto crash_loop_sleep_val
-              = config::node().crash_loop_sleep_sec.value();
-            if (crash_loop_sleep_val) {
-                vlog(
-                  _log.info,
-                  "Sleeping for {} seconds before terminating...",
-                  *crash_loop_sleep_val / 1s);
-                ss::sleep_abortable(*crash_loop_sleep_val, as).get();
-            }
-
-            throw std::runtime_error("Crash loop detected, aborting startup.");
-        }
-
-        vlog(
-          _log.debug,
-          "Consecutive crashes detected: {} node config changed: {} "
-          "time based tracking reset: {}",
-          crash_md._crash_count,
-          node_config_changed,
-          tracking_reset);
-
-        if (node_config_changed || tracking_reset) {
-            crash_md._crash_count = 0;
-        }
-    }
-
-    // Truncate and bump the crash count. We consider a run to be unclean by
-    // default unless the scheduled cleanup (that runs very late in shutdown)
-    // resets the file. See schedule_crash_tracker_file_cleanup().
-    auto new_crash_count = maybe_crash_md
-                             ? maybe_crash_md.value()._crash_count + 1
-                             : 1;
-    crash_tracker_metadata updated{
-      ._crash_count = new_crash_count,
-      ._config_checksum = checksum,
-      ._last_start_ts = model::timestamp::now()};
-    write_fully(file_path, serde::to_iobuf(updated)).get();
-    ss::sync_directory(config::node().data_directory.value().as_sstring())
-      .get();
+void application::init_crashtracker(::stop_signal& app_signal) {
+    _crash_tracker_service = std::make_unique<crash_tracker::service>();
+    _crash_tracker_service->start(app_signal.abort_source()).get();
 }
 
 void application::schedule_crash_tracker_file_cleanup() {
@@ -1138,12 +1034,8 @@ void application::schedule_crash_tracker_file_cleanup() {
     // next run.
     // We emplace it in the front to make it the last task to run.
     _deferred.emplace_front([&] {
-        auto file = config::node().crash_loop_tracker_path().string();
-        if (ss::file_exists(file).get()) {
-            ss::remove_file(file).get();
-            ss::sync_directory(config::node().data_directory().as_sstring())
-              .get();
-            vlog(_log.debug, "Deleted crash loop tracker file: {}", file);
+        if (_crash_tracker_service) {
+            _crash_tracker_service->stop().get();
         }
     });
 }
